Use tarsnap deb repository
Instead of compiling tarsnap from source, we use the deb repository for an
easier installation and upgrade process.  Now that it is a repository, it is
included inside `apt update`/`apt upgrade`, so we don't need to recompile for
updates.

A decision that I made was not to use the deb source installation method.  While
that is slightly more paranoid, I felt is was sufficient to just install the
normal deb repo.