My buster tweaks
The tweaks in this commit probably only apply to me.  Most of these just apply
to the way I particularly use Vagrant and sovereign server.

The most notable changes was the upgrade from using the
`debian/contrib-jessie64` vagrant box to using `generic/debian10`.  The
virtualization backend that I'm using is libvirt, and for some reason, the
official `debian/buster64` doesn't work for me (it hangs when it attempts to
connect to ssh).

I stopped running the restart ssh handler for Vagrant, since it would break when
sshd got restarted.

I also stopped running the tarsnap tasks for Vagrant, since I don't particularly
want to back up a virtual machine.