Update yunohost Rspamd repository for Buster
I'm not really sure if this is needed or not, as I'm not running this on an
armv7l server, but given that @lessless has done this change, it seems like a
good idea for me to do it too.

Looking back on things, it would have been better to just fork their repository
and just work on Buster support from there, but I did not see that they were
already working on Buster support, so I had to do a lot of work by myself.

This seems to be the only substantial difference between @lessless's commits and
mine, apart from the fact that they explicitly stated that the PHP version is
`php7.3` in apt, not just asking to install `php` and hoping for the best (what
I did).