Ansible-lint Buster changes
When I was updating things to work with buster, I forgot to ansible-lint things.
One thing that needed to be linted was to change the Tarsnap installation check,
which previously checked if Tarsnap was installed as a specific version.
However, I changed the installation process to use the deb repository instead of
building things from a tar'd/zip'd source, and the shell does not need to
compare string equality.

Since the new check for Tarsnap does not need any shell functionality (it
basically just sees if the tarsnap command is a thing), it should be `command`
rather than `shell`.